requirements.txt
asgiref==3.11.0
Django==5.2.9
django-filter==25.2
djangorestframework==3.16.1
pillow==12.0.0
python-decouple==3.8
sqlparse==0.5.4
typing_extensions==4.15.0
tzdata==2025.2
